Feat: drill into a run from the Activity monitor (M595)
The live monitor showed each run as a one-liner ("calling shell", "thinking ·
iter 2") but you couldn't see what it actually did. Now every row is expandable:
click it to drill into the full detail — status / model / iterations / tokens /
cost, every tool call with its policy verdict and output, and the final answer —
fetched from the journal for that correlation and folded live as the agent works.

The run-detail renderer + live loader is extracted to a shared
`components/RunDetail.tsx` (RunDetailCards + RunDetailLoader) and now used by BOTH
the Runs history view and the Activity monitor, so a run reads identically
wherever you open it — and the fetch/subscribe/merge logic lives in one place.
The Runs view is rewired onto it with no behaviour change.

Verified end-to-end in a real browser (Playwright) against the demo echo daemon:
a run triggered live in Activity expands to show status=completed, model=mock,
1 iter, "no tool calls", final answer "[echo] drill into me", and a raw-events
(6) toggle; the refactored Runs view expands the same run identically; light +
dark; zero console errors under the strict CSP.

Local validation (org Actions minutes still exhausted — CI fails at startup,
steps=0): tsc --noEmit ✓ · vitest 51 ✓ · vite build ✓ · go build ./kernel/webui ✓
· dist rotation only · go.mod/go.sum untouched.
